requirements.txt
feedparser>=6.0.0
feedgen>=0.9.0
loguru>=0.7.0
requests>=2.31.0
python-dateutil>=2.8.0
